FDG-6820 Add social share button and refactor existing component
