+++ src/components/social-share/social-share-dropdown/social-share-dropdown.tsx
@@ -1,19 +1,56 @@
 import {
   icon,
-  shareButton
+  shareButton,
+  dropdown
 } from './social-share-dropdown.module.scss';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {faShareNodes} from '@fortawesome/free-solid-svg-icons';
-import React, {FunctionComponent} from 'react';
+import React, {FunctionComponent, useState} from 'react';
 import {IconProp} from "@fortawesome/fontawesome-svg-core";
+import {SocialShareComponent} from "../social-share";
 
 const SocialShareDropdown: FunctionComponent = () => {
+  const [dropdownOpen, setDropdownOpen] = useState(false);
+
+  const onButtonClick = () => {
+    setDropdownOpen(!dropdownOpen);
+  }
+  const copy = {
+    title: 'test',
+    description: 'test',
+    body: 'test',
+    emailSubject: 'test',
+    emailBody: 'test',
+    url: 'test',
+    image: 'test',
+  }
+
+
+
   return (
     <>
-      <div className={shareButton}>
+      <div
+        className={shareButton}
+        role={'button'}
+        tabIndex={0}
+        onClick={onButtonClick}
+        onKeyPress={onButtonClick}
+      >
         <FontAwesomeIcon icon={faShareNodes as IconProp} className={icon} />
         <span>Share</span>
       </div>
+      { dropdownOpen && (
+        <div
+          className={dropdown}
+          data-testid={'share-dropdown'}
+        >
+          <SocialShareComponent
+            copy={copy}
+            pageName={'test'}
+            displayStyle={'list'}
+          />
+        </div>
+      )}
     </>
   );
 }
