+++ src/components/social-share/social-share.module.scss.d.ts
@@ -67,8 +67,11 @@ export const horizontalSocialShareContent: string;
 export const horizontalShareButton: string;
 export const horizontalShareButtonContainer: string;
 export const socialShareContent: string;
+export const listSocialShareContent: string;
 export const shareButton: string;
+export const listShareButton: string;
 export const shareButtonContent: string;
+export const listShareButtonContent: string;
 export const shareButtonText: string;
 export const facebookIcon: string;
 export const twitterIcon: string;
