+++ src/layouts/currency-exchange-rates-converter/index.tsx
@@ -68,13 +68,13 @@ const CurrencyExchangeRatesConverter: FunctionComponent = () => {
         </div>
         <span>
             1.00 US Dollar = 0.92 Euro Zone-Euro
-          </span>
+        </span>
         <span className={footer}>
             The Currency Exchange Rates Converter tool is driven by the Treasury Reporting Rates of
             Exchange dataset. This dataset is updated quarterly and covers the period from
             December 31, 2022 to Month, DD, YYYY. For more information and to see the full dataset,
             please visit the Treasury Reporting Rates of Exchange dataset page.
-          </span>
+        </span>
       </div>
     </SiteLayout>
   )
