+++ src/layouts/feature/feature.tsx
@@ -20,7 +20,7 @@ import {MDXRenderer} from "gatsby-plugin-mdx";
 import { MDXProvider } from "@mdx-js/react"
 import dsmComponents from "./dsm/dsm";
 import {format} from "date-fns";
-import SocialShare from "../explainer/social-share/social-share";
+import SocialShare from "../../components/social-share/social-share";
 import InsightsImage from "../../components/insights-image/insights-image";
 import {IDataset} from "../../models/IDataset";
 import ExplainerRelatedDatasets from
