+++ src/models/ISocialShareComponent.ts
@@ -1,0 +1,11 @@
+import {ISocialShareCopy} from "./ISocialShareCopy";
+
+export interface ISocialShareComponent {
+  copy: ISocialShareCopy,
+  emailSeparator?: string,
+
+  pageName: string,
+  width?: number,
+  horizontal?: boolean,
+  displayStyle?: string,
+}
