+++ src/models/ISocialShareCopy.ts
@@ -1,0 +1,9 @@
+export interface ISocialShareCopy {
+  title: string,
+  description: string,
+  body: string,
+  emailSubject: string,
+  emailBody: string,
+  url: string,
+  image: string,
+}
